Move all the agent run logic to the run command
This makes the 'start' command a pure alias
